chore: revert "feat: add redis support in shopify pixel for id stitching (#3957)"
This reverts commit 165e06de1ac22cda5c4db0b0b23e6e9972d5fa94.
