+++ src/v0/sources/shopify/util.js
@@ -272,7 +272,6 @@ module.exports = {
   createPropertiesForEcomEvent,
   extractEmailFromPayload,
   getAnonymousIdAndSessionId,
-  getCartToken,
   checkAndUpdateCartItems,
   getHashLineItems,
   getDataFromRedis,
