--- src/v1/sources/shopify/pixelEventsMappings/campaignObjectMappings.json
@@ -1,18 +1,0 @@
-[
-  {
-    "sourceKeys": "utm_campaign",
-    "destKeys": "name"
-  },
-  {
-    "sourceKeys": "utm_medium",
-    "destKeys": "medium"
-  },
-  {
-    "sourceKeys": "utm_term",
-    "destKeys": "term"
-  },
-  {
-    "sourceKeys": "utm_content",
-    "destKeys": "content"
-  }
-]
