+++ src/v1/sources/shopify/transform.js
@@ -4,14 +4,10 @@ const { process: processWebhookEvents } = require('../../../v0/sources/shopify/t
 const {
   process: processPixelWebhookEvents,
 } = require('./webhookTransformations/serverSideTransform');
-const { processIdentifierEvent, isIdentifierEvent } = require('./utils');
 
 const process = async (inputEvent) => {
   const { event } = inputEvent;
   const { query_parameters } = event;
-  if (isIdentifierEvent(event)) {
-    return processIdentifierEvent(event);
-  }
   // check identify the event is from the web pixel based on the pixelEventLabel property.
   const { pixelEventLabel: pixelClientEventLabel } = event;
   if (pixelClientEventLabel) {
