+++ test/integrations/sources/shopify/pixelTestScenarios/ProductEventsTests.ts
@@ -1,11 +1,5 @@
 // This file contains the test scenarios related to Shopify pixel events, emitted from web pixel on the browser.
-import {
-  dummyContext,
-  dummyContextwithCampaign,
-  dummySourceConfig,
-  responseDummyContext,
-  responseDummyContextwithCampaign,
-} from '../constants';
+import { dummyContext, dummySourceConfig, responseDummyContext } from '../constants';
 
 export const pixelEventsTestScenarios = [
   {
@@ -24,7 +18,7 @@ export const pixelEventsTestScenarios = [
               type: 'standard',
               clientId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
               timestamp: '2024-09-15T17:24:30.373Z',
-              context: dummyContextwithCampaign,
+              context: dummyContext,
               pixelEventLabel: true,
               query_parameters: {
                 topic: ['page_viewed'],
@@ -48,14 +42,7 @@ export const pixelEventsTestScenarios = [
               batch: [
                 {
                   context: {
-                    ...responseDummyContextwithCampaign,
-                    campaign: {
-                      content: 'web',
-                      medium: 'checkout',
-                      name: 'shopifySale',
-                      term: 'term_checkout',
-                      utm_custom1: 'customutm',
-                    },
+                    ...responseDummyContext,
                     shopifyDetails: {
                       clientId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
                       data: {},
